requirements.txt
streamlit==1.26.0
tensorflow==2.15.0
numpy==1.24.3
opencv-python-headless==4.8.0.74
Pillow==9.5.0
scikit-learn==1.3.0
matplotlib==3.7.1
pandas==1.5.3
joblib==1.3.2
scipy==1.10.1
gdown
